LayoutTests/imported/w3c/web-platform-tests/css/css-multicol/multicol-fill-balance-028-expected.txt


FAIL Make room for the float in the first column assert_equals: expected 100 but got 95
